Migrate all unit tests inline
